Add client api

+++ Common/include/s5_tcp_connection.h
@@ -27,7 +27,7 @@ public:
 	static void on_send_q_event(int fd, uint32_t event, void* c);
 	static void on_recv_q_event(int fd, uint32_t event, void* c);
 	static void on_socket_event(int fd, uint32_t event, void* c);
-	static S5TcpConnection* connect_to_server(const std::string& ip, short port, S5Poller *poller,
+	static S5TcpConnection* connect_to_server(const std::string& ip, int port, S5Poller *poller,
 		S5ClientVolumeInfo* vol, int timeout_sec);
 
 	int init(int sock_fd, S5Poller *poller, int send_q_depth, int recv_q_depth);
